client refactor
