+++ backend/models/Member.js
@@ -51,10 +51,6 @@ const memberSchema = new Schema ({
     type: Date, 
     default: Date.now()
   }
-
-})
-
-
-
+});
 
 module.exports = mongoose.model('Member', memberSchema);
