+++ frontend/src/App.js
@@ -1,23 +1,22 @@
 import {Routes, Route} from "react-router-dom"; 
-import Home from "./components/Home"; 
-import Signin from "./components/Signin"; 
-import Register from "./components/Register";
-import Family from "./components/Family";
-import Member from "./components/Member";
-import Dashboard from "./components/Dashboard";
+import Home from "./pages/Home"; 
+import Signin from "./pages/Signin"; 
+import Register from "./pages/Register";
+import Family from "./pages/Family";
+import Member from "./pages/Member";
+import Dashboard from "./pages/Dashboard";
 
 function App() {
   return (
     <div className="App">
       <Routes> 
         <Route> 
-          <Route path="/" element = {<Home />} /> 
-          <Route path="/signin" element = {<Signin />} /> 
-          <Route path="/register" element = {<Register />} /> 
-          <Route path="/family" element = {<Family />} /> 
-          <Route path="/member" element = {<Member />} /> 
-          <Route path="/dashboard" element = {<Dashboard />} /> 
-
+          <Route path="/" element={<Home />} /> 
+          <Route path="/signin" element={<Signin />} /> 
+          <Route path="/register" element={<Register />} /> 
+          <Route path="/family" element={<Family />} /> 
+          <Route path="/member" element={<Member />} /> 
+          <Route path="/dashboard" element={<Dashboard />} /> 
         </Route>
       </Routes>
     </div>
