+++ frontend/src/pages/Home/Home.css
@@ -177,7 +177,7 @@ body.home {
    ====================================== */
 
 .banner-section#header-banner {
-  background: #f8f8ee url('../media/homePage/kidsfirst_intro_1.png') no-repeat
+  background: #f8f8ee url('../../media/homePage/kidsfirst_intro_1.png') no-repeat
     right bottom;
   min-height: 730px;
   width: 100%;
@@ -197,7 +197,7 @@ body.home {
 }
 
 .banner-section#middle-banner {
-  background: #f8f8ee url('../media/homePage/kidsfirst_intro_2.png') no-repeat
+  background: #f8f8ee url('../../media/homePage/kidsfirst_intro_2.png') no-repeat
     left bottom;
   min-height: 450px;
   width: 100%;
