+++ frontend/src/pages/Home/Home.jsx
@@ -1,14 +1,14 @@
 import React from 'react'
 import { Button, Dropdown } from 'react-bootstrap'
-import './home.css'
-import logo from '../media/kids_first_logo_beta.png'
-import value_1 from '../media/homePage/value_1.png'
-import value_2 from '../media/homePage/value_2.png'
-import value_3 from '../media/homePage/value_3.png'
-import feature_1 from '../media/homePage/feature_1.png'
-import feature_2 from '../media/homePage/feature_2.png'
-import feature_3 from '../media/homePage/feature_3.png'
-import video from '../media/homePage/video_placeholder.png'
+import './Home.css'
+import logo from '../../media/kids_first_logo_beta.png'
+import value_1 from '../../media/homePage/value_1.png'
+import value_2 from '../../media/homePage/value_2.png'
+import value_3 from '../../media/homePage/value_3.png'
+import feature_1 from '../../media/homePage/feature_1.png'
+import feature_2 from '../../media/homePage/feature_2.png'
+import feature_3 from '../../media/homePage/feature_3.png'
+import video from '../../media/homePage/video_placeholder.png'
 
 import { NavLink } from 'react-router-dom'
