+++ frontend/src/pages/Signin/Signin.jsx
@@ -1,45 +1,30 @@
 import { useState } from "react";
-import axios from "axios";
 
 import Button from "react-bootstrap/Button";
 import Form from "react-bootstrap/Form";
 
-export default function Register() {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [error, setError] = useState("");
+import { login } from "../../api";
+
+export default function Signin() {
+  const [email, setEmail] = useState();
+  const [password, setPassword] = useState();
 
   console.log("email is ", email);
   console.log("password is ", password);
 
-  function handleRegister(e) {
+  function handleLogin(e) {
     e.preventDefault();
 
     console.log("email is ", email);
     console.log("password is ", password);
 
-    axios
-      .post("http://localhost:8000/api/register", { email, password })
-      // next page should - give a family name
-      .then((res) => {
-        console.log("success");
-        // need to store the user informatin in the session 
-        const user = JSON.stringify(res.data);
-        localStorage.setItem("storedUser", user); 
-        window.location.href = "/family";
-      })
-      .catch((e) => {
-        console.log(e.response.data.message);
-        setError(e.response.data.message);
-      });
-    // then firstname, last name ... fill out the information .
+    login(email, password).then(console.log("posted"));
   }
 
   return (
     <div>
-      registration page
-      {error.length > 1 ? error : ""}
-      <Form onSubmit={handleRegister}>
+      sign in page
+      <Form onSubmit={handleLogin}>
         <Form.Group className="mb-3" controlId="formBasicEmail">
           <Form.Label>Email address</Form.Label>
           <Form.Control
@@ -59,7 +44,7 @@ export default function Register() {
         </Form.Group>
 
         <Button variant="primary" type="submit">
-          Register
+          Sign in
         </Button>
       </Form>
     </div>
