+++ frontend/src/pages/Signin/index.js
@@ -1,0 +1,3 @@
+import Signin from "./Signin";
+
+export default Signin;
